% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_table.R
\name{cross_tbl}
\alias{cross_tbl}
\title{cross_tbl}
\usage{
cross_tbl(obj, var1, var2, title = "", marg = FALSE)
}
\arguments{
\item{obj}{Object used to create the table. Data frame, list, or environment
(or object coercible by as.data.frame to a data frame)}

\item{var1}{Variable that you want the table (not written in string format)}

\item{var2}{Variable that you want on the top of the table (not written in string format)}

\item{title}{title for the table, write in string format}

\item{marg}{Marginal row table, default is FALSE}
}
\value{
A knitr_kable with 2x2 table fitted
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function makes cross tables, like all functions from fastrep
you need to supply a data.frame, and in this case two variables.
}
\examples{
mtcars |>
  fastrep::cross_tbl(cyl, am, "title", marg = TRUE)

}
