% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq-table.R
\name{freq_tbl}
\alias{freq_tbl}
\title{freq_tbl}
\usage{
freq_tbl(
  obj,
  var,
  sort_by = {
     {
         var
     }
 },
  desc = FALSE
)
}
\arguments{
\item{obj}{Object used to create the table. Data frame, list or environment
(or object coercible by as.data.frame to a data frame)}

\item{var}{Variable that you want the table (not written in string format)}

\item{sort_by}{Variable you want to sort the \code{tibble}}

\item{desc}{if you want decreasing order put \code{FALSE}}
}
\value{
A tibble with 4 rows and 3 columns
}
\description{
This function creates a frequency table, you only need to supply a
data.frame and the variable to make the table.
}
\note{
By default the \code{sort_by} is in alphabetical order of  \code{var}
}
\examples{
iris \%>\%
  fastrep::freq_tbl(Species) \%>\%
  fastrep::tbl("You can combine this function too!")

}
