% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{pdf1_citation}
\alias{pdf1_citation}
\title{citation
How to cite \code{mypdf1} in your publications!}
\usage{
pdf1_citation()
}
\value{
A print in console indicating how to cite the package.
}
\description{
citation
How to cite \code{mypdf1} in your publications!
}
