% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq-table-DE.R
\name{pdf1_freq.tbl}
\alias{pdf1_freq.tbl}
\title{pdf1_freq.tbl}
\usage{
pdf1_freq.tbl(obj, v, tit)
}
\arguments{
\item{obj}{Object used to create the table. Data frame, list or environment
(or object coercible by as.data.frame to a data frame)}

\item{v}{Variable that you want the table (not written in string format)}

\item{tit}{Title for the table, write in string format}
}
\description{
this is a very simple frequency table generator
}
