% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf1_na.R
\name{pdf1_na}
\alias{pdf1_na}
\title{pdf1_na}
\usage{
pdf1_na(obj, ...)
}
\arguments{
\item{obj}{Object used to create the table.

\code{data.frame}, \code{list} or environment
(or object coercible by \code{as.data.frame} to a \code{data.frame})}

\item{...}{Other arguments}
}
\value{
A tibble with n rows where n is equal to \code{ncol(obj)} and 2 columns
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function counts the number of NA by variable in your data.frame
}
\examples{
airquality \%>\%
  mypdf1:::pdf1_na()
}
