% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf1_summary.R
\name{pdf1_summary}
\alias{pdf1_summary}
\title{pdf1_summary()}
\usage{
pdf1_summary(obj, na_rm = TRUE, use_num = FALSE)
}
\arguments{
\item{obj}{Object used to create the table. Data frame, list or environment
(or object coercible by as.data.frame to a data frame)}

\item{na_rm}{option to remove NA from variables}

\item{use_num}{option to use only numeric variables}
}
\description{
this is a very summary  generator
}
\examples{
mtcars |>
  mypdf1::pdf1_summary()

airquality |> mypdf1::pdf1_summary(na_rm = FALSE)

iris |> mypdf1::pdf1_summary(use_num = TRUE)
}
