% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf1_tbl.R
\name{pdf1_tbl}
\alias{pdf1_tbl}
\title{pdf1_tbl}
\usage{
pdf1_tbl(v, tit)
}
\arguments{
\item{v}{Objeto a ser printado}

\item{tit}{titulo da tabela}
}
\value{
não sei
}
\description{
function to print  a table
}
