% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq-table.R
\name{pdf1_tbl_freq}
\alias{pdf1_tbl_freq}
\title{pdf1_freq_tbl}
\usage{
pdf1_tbl_freq(obj, var)
}
\arguments{
\item{obj}{Object used to create the table. Data frame, list or environment
(or object coercible by as.data.frame to a data frame)}

\item{var}{Variable that you want the table (not written in string format)}
}
\description{
this is a very simple frequency table generator
}
\examples{
iris \%>\%
  mypdf1::pdf1_tbl_freq(Species) \%>\%
  mypdf1::pdf1_tbl(" You can combine this function too!")
}
