% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{rep_citation}
\alias{rep_citation}
\title{citation
How to cite \code{fastrep} in your publications!}
\usage{
rep_citation()
}
\value{
A print in console indicating how to cite the package.
}
\description{
citation
How to cite \code{fastrep} in your publications!
}
