% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{tab}
\usage{
tab(obj, var1, var2)
}
\arguments{
\item{obj}{Object used to create the table. Data frame, list, or environment
(or object coercible by as.data.frame to a data frame)}

\item{var1}{Variable that you want the table (not written in string format)}

\item{var2}{Variable that you want on the top of the table (not written in string format)}
}
\value{
A tibble to create \code{pdf1_cross_tbl}
}
\description{
This function is used to create \code{pdf1_cross_tbl}
}
\examples{
mtcars |>
  mypdf1::pdf1_cross_tbl(cyl, am)

}
