% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf1_na.R
\name{pdf1_na}
\alias{pdf1_na}
\title{pdf1_na}
\usage{
pdf1_na(obj, ...)
}
\arguments{
\item{obj}{Object used to create the table.

\code{data.frame}, \code{list} or environment
(or object coercible by \code{as.data.frame} to a \code{data.frame})}

\item{...}{Other arguments}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

this is a very simple table of the quantity of \code{NA} by variable
}
\examples{
airquality \%>\%
  mypdf1::pdf1_na()
}
