% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf1_na.R
\name{pdf1_na}
\alias{pdf1_na}
\title{pdf1_na}
\usage{
pdf1_na(obj, ...)
}
\arguments{
\item{obj}{Object used to create the table. Data frame, list or environment
(or object coercible by as.data.frame to a data frame)}

\item{...}{Other arguments}
}
\description{
this is a very simple frequency table generator
}
\examples{
airquality \%>\%
mypdf1::pdf1_na()
}
