% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf1_summary.R
\name{pdf1_summary}
\alias{pdf1_summary}
\title{pdf1_summary()}
\usage{
pdf1_summary(obj, na_rm = TRUE)
}
\arguments{
\item{obj}{Object used to create the table.
\code{data.frame}, \code{list} or environment
(or object coercible by \code{as.data.frame} to a \code{data.frame})}

\item{na_rm}{option to remove \code{NA} from variables}
}
\value{
A tibble with n rows where n is equal to \code{ncol(obj)} and  columns
with
the summary metrics
}
\description{
In the base R we have the function summary, but the output is no by default
a \code{data.frame}, so \code{pdf1_summary} is an enhancement of this function
to summarize data frames.
}
\examples{
mtcars |>
  mypdf1::pdf1_summary()

airquality |> mypdf1::pdf1_summary(na_rm = FALSE)

iris |> mypdf1::pdf1_summary()

}
