% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf1_tbl.R
\name{pdf1_tbl}
\alias{pdf1_tbl}
\title{pdf1_tbl}
\usage{
pdf1_tbl(obj, tit, ...)
}
\arguments{
\item{obj}{Object used to create the table}

\item{tit}{Title for the table, write in string format}

\item{...}{Other arguments}
}
\description{
this is a very simple table generator
}
