% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq-table.R
\name{pdf1_tbl_freq}
\alias{pdf1_tbl_freq}
\title{pdf1_freq_tbl}
\usage{
pdf1_tbl_freq(
  obj,
  var,
  sort_by = {
     {
         var
     }
 },
  desc = F
)
}
\arguments{
\item{obj}{Object used to create the table. Data frame, list or environment
(or object coercible by as.data.frame to a data frame)}

\item{var}{Variable that you want the table (not written in string format)}

\item{sort_by}{Variable you want to sort the \code{tibble}}

\item{desc}{if you want decreasing order put \code{FALSE}}
}
\value{
A tibble with 4 rows and 3 columns
}
\description{
This function creates a frequency table, you only need to supply a
data.frame and the variable to make the table.
}
\note{
By default the \code{sort_by} is in alphabetical order of \code{{{ var }}}
}
\examples{
iris \%>\%
  mypdf1::pdf1_tbl_freq(Species) \%>\%
  mypdf1::pdf1_tbl("You can combine this function too!")

}
